requirements.txt
packaging
timm==0.4.12
pytest 
chardet 
yacs 
termcolor
submitit 
tensorboardX
fvcore
seaborn
ipdb==0.13.11
albumentations==1.2.1
torchmetrics==0.9.3
Pillow==9.1.1
tqdm==4.64.0
wandb==0.12.21
pathlib==1.0.1
opencv-python==4.6.0.66
numpy==1.22.4